Set aria-hidden on outside elements for popovers and fix focusing dialogs in iOS VoiceOver
