+++ packages/@react-aria/dialog/src/index.ts
@@ -11,4 +11,3 @@
  */
 
 export * from './useDialog';
-export * from './useModalDialog';
