+++ packages/@react-aria/overlays/src/usePreventScroll.ts
@@ -12,7 +12,7 @@
 
 import {useEffect} from 'react';
 
-export function useModal() {
+export function usePreventScroll() {
   // Add overflow: hidden to the body on mount, and restore on unmount.
   useEffect(() => {
     let overflow = document.body.style.overflow;
